requirements.txt
# Core dependencies
python-dotenv
requests
pandas
numpy

# Visualization (used during data exploration)
matplotlib
streamlit
plotly

# Machine learning utilities
scikit-learn
joblib
xgboost

# Hopsworks Feature Store SDK
hopsworks==4.2.2
hopsworks[python]==4.2.2 
confluent-kafka